+++ src/TitleDetails/TitleInformation/TitleInformation.js
@@ -1,7 +1,6 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
 
 import {
   Col,
@@ -18,55 +17,16 @@ import {
 const TitleInformation = ({
   contributors,
   edition,
-  instanceId,
-  poLineId,
-  poLineNumber,
   productIds,
   publishedDate,
   publisher,
-  receiptDate,
-  receivingNote,
   subscriptionFrom,
   subscriptionInterval,
   subscriptionTo,
-  title,
 }) => {
-  const titleValue = instanceId
-    ? (
-      <Link
-        data-testid="titleInstanceLink"
-        to={`/inventory/view/${instanceId}`}
-      >
-        {title}
-      </Link>
-    )
-    : title;
-
-  const poLineNumberValue = (
-    <Link
-      data-testid="titlePOLineLink"
-      to={{
-        pathname: `/orders/lines/view/${poLineId}`,
-        search: `?qindex=poLineNumber&query=${poLineNumber}`,
-      }}
-    >
-      {poLineNumber}
-    </Link>
-  );
-
   return (
-    <Fragment>
+    <>
       <Row>
-        <Col
-          data-test-title-information-title
-          xs={6}
-          md={3}
-        >
-          <KeyValue
-            label={<FormattedMessage id="ui-receiving.title.title" />}
-            value={titleValue}
-          />
-        </Col>
         <Col
           data-test-title-information-publisher
           xs={6}
@@ -148,73 +108,30 @@ const TitleInformation = ({
           </KeyValue>
         </Col>
       </Row>
-      <Row>
-        <Col
-          data-test-title-po-line-number
-          xs={6}
-          md={3}
-        >
-          <KeyValue
-            label={<FormattedMessage id="ui-receiving.title.polNumber" />}
-            value={poLineNumberValue}
-          />
-        </Col>
-        <Col
-          data-test-title-receipt-date
-          xs={6}
-          md={3}
-        >
-          <KeyValue
-            label={<FormattedMessage id="ui-receiving.title.expectedReceiptDate" />}
-            value={<FolioFormattedDate value={receiptDate} />}
-          />
-        </Col>
-        <Col
-          data-test-title-receiving-note
-          xs={6}
-          md={3}
-        >
-          <KeyValue
-            label={<FormattedMessage id="ui-receiving.title.receivingNote" />}
-            value={receivingNote}
-          />
-        </Col>
-      </Row>
-    </Fragment>
+    </>
   );
 };
 
 TitleInformation.propTypes = {
   contributors: PropTypes.arrayOf(PropTypes.object),
   edition: PropTypes.string,
-  instanceId: PropTypes.string,
-  poLineId: PropTypes.string,
-  poLineNumber: PropTypes.string,
   productIds: PropTypes.arrayOf(PropTypes.object),
   publishedDate: PropTypes.string,
   publisher: PropTypes.string,
-  receiptDate: PropTypes.string,
-  receivingNote: PropTypes.string,
   subscriptionFrom: PropTypes.string,
   subscriptionInterval: PropTypes.number,
   subscriptionTo: PropTypes.string,
-  title: PropTypes.string,
 };
 
 TitleInformation.defaultProps = {
   contributors: [],
   edition: '',
-  instanceId: '',
-  poLineNumber: '',
   productIds: [],
   publishedDate: '',
   publisher: '',
-  receiptDate: '',
-  receivingNote: '',
   subscriptionFrom: '',
   subscriptionInterval: 0,
   subscriptionTo: '',
-  title: '',
 };
 
 export default TitleInformation;