+++ test/bigtest/tests/title/title-details-test.js
@@ -1,7 +1,10 @@
 import { beforeEach, describe, it } from '@bigtest/mocha';
 import { expect } from 'chai';
 
-import { ORDER_FORMATS } from '@folio/stripes-acq-components';
+import {
+  ORDER_FORMATS,
+  ORDER_STATUSES,
+} from '@folio/stripes-acq-components';
 
 import { PIECE_FORMAT } from '../../../../src/common/constants';
 import setupApplication from '../../helpers/setup-application';
@@ -17,8 +20,15 @@ describe('Title details', () => {
   setupApplication();
 
   beforeEach(async function () {
+    const order = this.server.create('order', {
+      workflowStatus: ORDER_STATUSES.closed,
+      closeReason: {
+        reason: 'Reason for closure',
+      },
+    });
     const line = this.server.create('line', {
       orderFormat: ORDER_FORMATS.physicalResource,
+      purchaseOrderId: order.id,
     });
     const title = this.server.create('title', {
       poLineId: line.id,
@@ -38,6 +48,10 @@ describe('Title details', () => {
     expect(titleDetails.receiveButton.isPresent).to.be.true;
   });
 
+  it('warning message with closing reason is shown', () => {
+    expect(titleDetails.closingReasonMessage).to.be.true;
+  });
+
   describe('click receive button', function () {
     beforeEach(async function () {
       await titleDetails.receiveButton.click();
@@ -46,6 +60,26 @@ describe('Title details', () => {
 
     it('shows Title receive screen', function () {
       expect(titleReceive.isPresent).to.be.true;
+    });
+  });
+
+  describe('click on title', function () {
+    beforeEach(async function () {
+      await titleDetails.instanceLink();
+    });
+
+    it('goes to instance in inventory app', function () {
+      expect(titleReceive.isPresent).to.be.false;
+    });
+  });
+
+  describe('click on PO number', function () {
+    beforeEach(async function () {
+      await titleDetails.poLineLink();
+    });
+
+    it('goes to POL in orders app', function () {
+      expect(titleReceive.isPresent).to.be.false;
     });
   });
 });